refactor: add timestamp header
