+++ rust/batch-import-worker/src/parse/content/amplitude/identify.rs
@@ -74,6 +74,7 @@ pub fn create_identify_event(
         sent_at: None,
         token: token.to_string(),
         event: "$identify".to_string(),
+        timestamp,
         is_cookieless_mode: false,
         historical_migration: true,
     };
