+++ rust/capture/src/ai_endpoint.rs
@@ -268,6 +268,7 @@ fn build_kafka_event(
         sent_at: parsed.sent_at,
         token: token.to_string(),
         event: parsed.event_name.clone(),
+        timestamp: computed_timestamp,
         is_cookieless_mode: false,
         historical_migration: false,
     };
