+++ rust/capture/src/lib.rs
@@ -9,8 +9,10 @@ pub mod server;
 pub mod sinks;
 pub mod test_endpoint;
 pub mod time;
-pub mod timestamp;
 pub mod token;
 pub mod utils;
 pub mod v0_endpoint;
 pub mod v0_request;
+
+// Re-export timestamp parsing from common-types for backwards compatibility
+pub use common_types::timestamp;
