--- rust/capture/src/timestamp.rs
@@ -1,721 +1,0 @@
-use std::borrow::Cow;
-
-use chrono::{DateTime, Datelike, Duration, Utc};
-use jiff::civil::DateTime as JiffDateTime;
-use regex::Regex;
-
-const FUTURE_EVENT_HOURS_CUTOFF_MILLIS: i64 = 23 * 3600 * 1000; // 23 hours
-
-/// Parse event timestamp with clock skew adjustment and validation
-///
-/// # Arguments
-/// * `timestamp` - The event timestamp string (optional)
-/// * `offset` - The offset in milliseconds (optional)
-/// * `sent_at` - The client-sent timestamp (optional)
-/// * `ignore_sent_at` - Whether to ignore sent_at for clock skew adjustment
-/// * `now` - The current server timestamp
-///
-/// # Returns
-/// * `DateTime<Utc>` - The parsed and validated timestamp
-pub fn parse_event_timestamp(
-    timestamp: Option<&str>,
-    offset: Option<i64>,
-    sent_at: Option<DateTime<Utc>>,
-    ignore_sent_at: bool,
-    now: DateTime<Utc>,
-) -> DateTime<Utc> {
-    // Use sent_at only if not ignored
-    let effective_sent_at = if ignore_sent_at { None } else { sent_at };
-
-    // Handle timestamp parsing and clock skew adjustment
-    let mut parsed_ts = handle_timestamp(timestamp, offset, effective_sent_at, now);
-
-    // Check for future events - clamp to now
-    let now_diff = parsed_ts.signed_duration_since(now).num_milliseconds();
-    if now_diff > FUTURE_EVENT_HOURS_CUTOFF_MILLIS {
-        parsed_ts = now;
-    }
-
-    // Check if timestamp is out of bounds - fallback to epoch
-    if parsed_ts.year() < 0 || parsed_ts.year() > 9999 {
-        parsed_ts = DateTime::UNIX_EPOCH;
-    }
-
-    parsed_ts
-}
-
-fn handle_timestamp(
-    timestamp: Option<&str>,
-    offset: Option<i64>,
-    sent_at: Option<DateTime<Utc>>,
-    now: DateTime<Utc>,
-) -> DateTime<Utc> {
-    let mut parsed_ts = now;
-
-    if let Some(timestamp_str) = timestamp {
-        let timestamp_parsed = parse_date(timestamp_str);
-
-        if let (Some(sent_at), Some(timestamp_parsed)) = (sent_at, timestamp_parsed) {
-            // Handle clock skew between client and server
-            // skew = sent_at - now
-            // x = now + (timestamp - sent_at)
-            let duration = timestamp_parsed.signed_duration_since(sent_at);
-            parsed_ts = now + duration;
-        } else if let Some(timestamp_parsed) = timestamp_parsed {
-            parsed_ts = timestamp_parsed;
-        }
-    }
-
-    // Handle offset if present
-    if let Some(offset_ms) = offset {
-        parsed_ts = now - Duration::milliseconds(offset_ms);
-    }
-
-    parsed_ts
-}
-
-/// Parse a date string using a streamlined approach
-///
-/// This function tries parsing in order of preference:
-/// 1. chrono RFC3339 parser (handles standard ISO 8601 with proper timezone conversion)
-/// 2. dateparser (handles 95%+ of formats): ISO 8601, slash-separated, RFC2822, numeric timestamps
-/// 3. jiff (minimal fallback): civil datetime with T but no timezone (e.g., "2023-01-01T12:00:00")
-fn parse_date(supposed_iso_string: &str) -> Option<DateTime<Utc>> {
-    // First normalize any non-standard timezone formats (e.g., +03 -> +03:00)
-    let normalized_input = normalize_timezone_format(supposed_iso_string);
-
-    // Try chrono's RFC3339 parser first for proper timezone handling
-    if let Ok(dt) = DateTime::parse_from_rfc3339(&normalized_input) {
-        return Some(dt.with_timezone(&Utc));
-    }
-
-    // Try dateparser for other formats - but note it may not handle timezones correctly
-    // - Date-only: 2023-01-01
-    // - Civil datetime with space: 2023-01-01 12:00:00
-    // - Slash-separated: 01/01/2023, 2023/01/01
-    // - RFC2822: Tue, 1 Jul 2003 10:52:37 +0200
-    // - Numeric timestamps: 1672574400000, 1672574400
-    if let Ok(dt) = dateparser::parse(&normalized_input) {
-        return Some(dt);
-    }
-
-    // Minimal jiff fallback for the one format dateparser can't handle:
-    // Civil datetime with T but no timezone (e.g., "2023-01-01T12:00:00")
-    if let Ok(jiff_civil) = normalized_input.parse::<JiffDateTime>() {
-        return convert_jiff_to_chrono(jiff_civil.to_zoned(jiff::tz::TimeZone::UTC).ok()?);
-    }
-
-    None
-}
-
-/// Normalize non-standard timezone formats to standard RFC3339 format
-/// Returns a Cow that borrows the input if no normalization is needed, or owns a new string if modified
-/// Uses regex to precisely match ISO datetime strings with non-standard timezone format
-/// Examples:
-/// - "2025-09-17T14:05:04.805+03" -> "2025-09-17T14:05:04.805+03:00" (owned)
-/// - "2025-09-17T14:05:04.805-05" -> "2025-09-17T14:05:04.805-05:00" (owned)
-/// - "2025-09-17T14:05:04.805Z" -> "2025-09-17T14:05:04.805Z" (borrowed)
-/// - "2023-01-01" -> "2023-01-01" (borrowed, no match)
-fn normalize_timezone_format(input: &str) -> Cow<'_, str> {
-    // Quick optimization: check last 3 chars first
-    if input.len() < 3 {
-        return Cow::Borrowed(input);
-    }
-
-    let last_3_chars = &input[input.len() - 3..];
-    if !(last_3_chars.starts_with('+') || last_3_chars.starts_with('-'))
-        || !last_3_chars[1..].chars().all(|c| c.is_ascii_digit())
-    {
-        return Cow::Borrowed(input);
-    }
-
-    // Use regex to confirm this is an ISO datetime with non-standard timezone format
-    // Pattern: YYYY-MM-DDTHH:MM:SS[.fff][+/-]HH
-    static TIMEZONE_REGEX: std::sync::LazyLock<Regex> = std::sync::LazyLock::new(|| {
-        Regex::new(r"^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(?:\.\d{1,9})?[+-]\d{2}$").unwrap()
-    });
-
-    if TIMEZONE_REGEX.is_match(input) {
-        // Found ISO datetime with +XX or -XX timezone, convert to +XX:00 or -XX:00
-        Cow::Owned(format!("{input}:00"))
-    } else {
-        // Not the format we're looking for, return unchanged
-        Cow::Borrowed(input)
-    }
-}
-
-/// Helper function to convert jiff timestamp to chrono DateTime<Utc>
-fn convert_jiff_to_chrono(jiff_timestamp: jiff::Zoned) -> Option<DateTime<Utc>> {
-    let seconds = jiff_timestamp.timestamp().as_second();
-    let nanos = jiff_timestamp.timestamp().subsec_nanosecond();
-    // Convert i32 to u32 safely (nanoseconds should always be positive)
-    let nanos_u32 = nanos.try_into().ok()?;
-    DateTime::from_timestamp(seconds, nanos_u32)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use chrono::{TimeZone, Timelike, Utc};
-
-    #[test]
-    fn test_parse_event_timestamp_basic() {
-        let now_str = "2023-01-01T12:00:00Z";
-        let now = DateTime::parse_from_rfc3339(now_str)
-            .unwrap()
-            .with_timezone(&Utc);
-
-        let result = parse_event_timestamp(None, None, None, false, now);
-
-        assert_eq!(result, now);
-    }
-
-    #[test]
-    fn test_parse_event_timestamp_with_clock_skew() {
-        let now_str = "2023-01-01T12:00:00Z";
-        let now = DateTime::parse_from_rfc3339(now_str)
-            .unwrap()
-            .with_timezone(&Utc);
-        let sent_at_str = "2023-01-01T12:00:05Z"; // 5 seconds ahead
-        let timestamp_str = "2023-01-01T11:59:55Z"; // 10 seconds before sent_at
-        let sent_at = DateTime::parse_from_rfc3339(sent_at_str)
-            .unwrap()
-            .with_timezone(&Utc);
-
-        let result = parse_event_timestamp(Some(timestamp_str), None, Some(sent_at), false, now);
-        // Expected: now + (timestamp - sent_at) = 12:00:00 + (11:59:55 - 12:00:05) = 12:00:00 - 00:00:10 = 11:59:50
-        let expected = Utc.with_ymd_and_hms(2023, 1, 1, 11, 59, 50).unwrap();
-
-        assert_eq!(result, expected);
-    }
-
-    #[test]
-    fn test_parse_event_timestamp_with_offset() {
-        let now_str = "2023-01-01T12:00:00Z";
-        let now = DateTime::parse_from_rfc3339(now_str)
-            .unwrap()
-            .with_timezone(&Utc);
-        let offset = 5000; // 5 seconds
-
-        let result = parse_event_timestamp(None, Some(offset), None, false, now);
-        let expected = Utc.with_ymd_and_hms(2023, 1, 1, 11, 59, 55).unwrap();
-
-        assert_eq!(result, expected);
-    }
-
-    #[test]
-    fn test_parse_event_timestamp_ignore_sent_at() {
-        let now_str = "2023-01-01T12:00:00Z";
-        let now = DateTime::parse_from_rfc3339(now_str)
-            .unwrap()
-            .with_timezone(&Utc);
-        let sent_at_str = "2023-01-01T12:00:05Z";
-        let timestamp_str = "2023-01-01T11:00:00Z";
-        let sent_at = DateTime::parse_from_rfc3339(sent_at_str)
-            .unwrap()
-            .with_timezone(&Utc);
-
-        let result = parse_event_timestamp(Some(timestamp_str), None, Some(sent_at), true, now);
-        // Should use timestamp directly since sent_at is ignored
-        let expected = Utc.with_ymd_and_hms(2023, 1, 1, 11, 0, 0).unwrap();
-
-        assert_eq!(result, expected);
-    }
-
-    #[test]
-    fn test_parse_event_timestamp_future_event() {
-        let now_str = "2023-01-01T12:00:00Z";
-        let now = DateTime::parse_from_rfc3339(now_str)
-            .unwrap()
-            .with_timezone(&Utc);
-        let future_timestamp = "2023-01-02T12:00:00Z"; // 24 hours in the future
-
-        let result = parse_event_timestamp(Some(future_timestamp), None, None, false, now);
-
-        // Should clamp to now for future events
-        let expected = DateTime::parse_from_rfc3339(now_str)
-            .unwrap()
-            .with_timezone(&Utc);
-        assert_eq!(result, expected);
-    }
-
-    #[test]
-    fn test_parse_event_timestamp_with_millisecond_precision() {
-        let now_str = "2023-01-01T12:00:00.100Z";
-        let now = DateTime::parse_from_rfc3339(now_str)
-            .unwrap()
-            .with_timezone(&Utc);
-        let sent_at_str = "2023-01-01T12:00:00.200Z"; // Client sent at 200ms
-        let timestamp_str = "2023-01-01T12:00:00.750Z"; // Event timestamp with 750ms
-        let sent_at = DateTime::parse_from_rfc3339(sent_at_str)
-            .unwrap()
-            .with_timezone(&Utc);
-
-        let result = parse_event_timestamp(Some(timestamp_str), None, Some(sent_at), false, now);
-
-        // Expected: now + (timestamp - sent_at)
-        // = 12:00:00.100 + (12:00:00.750 - 12:00:00.200)
-        // = 12:00:00.100 + 00:00:00.550
-        // = 12:00:00.650
-        let expected = Utc.with_ymd_and_hms(2023, 1, 1, 12, 0, 0).unwrap()
-            + chrono::Duration::milliseconds(650);
-
-        assert_eq!(result, expected);
-
-        // Verify millisecond precision is preserved
-        assert_eq!(result.timestamp_subsec_millis(), 650);
-    }
-
-    #[test]
-    fn test_parse_event_timestamp_out_of_bounds() {
-        let now_str = "2023-01-01T12:00:00Z";
-        let now = DateTime::parse_from_rfc3339(now_str)
-            .unwrap()
-            .with_timezone(&Utc);
-        let invalid_timestamp = "0001-01-01T12:00:00Z"; // Year 1 is within bounds, this will be parsed successfully
-
-        let result = parse_event_timestamp(Some(invalid_timestamp), None, None, false, now);
-
-        // Should use the parsed timestamp (year 1 is valid)
-        let expected = DateTime::parse_from_rfc3339(invalid_timestamp)
-            .unwrap()
-            .with_timezone(&Utc);
-        assert_eq!(result, expected);
-    }
-
-    #[test]
-    fn test_parse_event_timestamp_unparseable() {
-        let now_str = "2023-01-01T12:00:00Z";
-        let now = DateTime::parse_from_rfc3339(now_str)
-            .unwrap()
-            .with_timezone(&Utc);
-        let invalid_timestamp = "99999-01-01T12:00:00Z"; // This should fail to parse due to year being too large
-
-        let result = parse_event_timestamp(Some(invalid_timestamp), None, None, false, now);
-
-        // Should fall back to 'now' when timestamp fails to parse
-        let expected = DateTime::parse_from_rfc3339(now_str)
-            .unwrap()
-            .with_timezone(&Utc);
-        assert_eq!(result, expected);
-    }
-
-    #[test]
-    fn test_parse_date_various_formats() {
-        // Test core date parsing functionality across different format types
-        let valid_cases = vec![
-            // ISO 8601/RFC3339 (dateparser)
-            "2023-01-01T12:00:00Z",
-            "2023-01-01T12:00:00+02:00",
-            "2023-01-01",
-            // Civil datetime (jiff fallback)
-            "2023-01-01T12:00:00",
-            // Slash-separated (dateparser)
-            "01/01/2023",
-            "2023/01/01",
-            // RFC2822 (dateparser)
-            "Tue, 1 Jul 2003 10:52:37 +0200",
-            // Numeric timestamps (dateparser)
-            "1672574400000",
-            "1672574400",
-        ];
-
-        let invalid_cases = vec![
-            "invalid-date",
-            "99999-01-01T12:00:00Z", // Year too large
-            "13/32/2023",            // Invalid month/day
-            "",
-        ];
-
-        for date_str in valid_cases {
-            assert!(
-                parse_date(date_str).is_some(),
-                "Failed to parse: {date_str}"
-            );
-        }
-
-        for date_str in invalid_cases {
-            assert!(
-                parse_date(date_str).is_none(),
-                "Unexpectedly parsed: {date_str}"
-            );
-        }
-    }
-
-    #[test]
-    fn test_complex_timestamp_scenario() {
-        // Test complex scenario with all components
-        let now_str = "2023-01-01T12:00:00Z";
-        let now = DateTime::parse_from_rfc3339(now_str)
-            .unwrap()
-            .with_timezone(&Utc);
-        let sent_at_str = "2023-01-01T12:00:02Z"; // 2 seconds ahead of server
-        let timestamp_str = "2023-01-01T11:59:58Z"; // 4 seconds before sent_at
-        let sent_at = DateTime::parse_from_rfc3339(sent_at_str)
-            .unwrap()
-            .with_timezone(&Utc);
-
-        let result = parse_event_timestamp(Some(timestamp_str), None, Some(sent_at), false, now);
-
-        // Expected calculation:
-        // skew = sent_at - now = 12:00:02 - 12:00:00 = +2s
-        // timestamp_diff = timestamp - sent_at = 11:59:58 - 12:00:02 = -4s
-        // result = now + timestamp_diff = 12:00:00 + (-4s) = 11:59:56
-        let expected = Utc.with_ymd_and_hms(2023, 1, 1, 11, 59, 56).unwrap();
-
-        assert_eq!(result, expected);
-    }
-
-    // === Node.js Parity Tests ===
-    // Based on plugin-server/tests/worker/ingestion/timestamps.test.ts
-
-    #[test]
-    fn test_sent_at_timestamp_adjustment() {
-        // Matches: "captures sent_at to adjusts timestamp"
-        let now_str = "2021-10-29T01:44:00.000Z";
-        let now = DateTime::parse_from_rfc3339(now_str)
-            .unwrap()
-            .with_timezone(&Utc);
-        let timestamp_str = "2021-10-30T03:02:00.000Z";
-        let sent_at_str = "2021-10-30T03:12:00.000Z"; // 10 minutes ahead of timestamp
-        let sent_at = DateTime::parse_from_rfc3339(sent_at_str)
-            .unwrap()
-            .with_timezone(&Utc);
-
-        let result = parse_event_timestamp(Some(timestamp_str), None, Some(sent_at), false, now);
-
-        // Expected: now + (timestamp - sent_at) = 01:44:00 + (03:02:00 - 03:12:00) = 01:44:00 - 00:10:00 = 01:34:00
-        let expected = Utc.with_ymd_and_hms(2021, 10, 29, 1, 34, 0).unwrap();
-        assert_eq!(result, expected);
-    }
-
-    #[test]
-    fn test_ignore_sent_at_property() {
-        // Matches: "Ignores sent_at if $ignore_sent_at set"
-        let now_str = "2021-11-29T01:44:00.000Z";
-        let now = DateTime::parse_from_rfc3339(now_str)
-            .unwrap()
-            .with_timezone(&Utc);
-        let timestamp_str = "2021-10-30T03:02:00.000Z";
-        let sent_at_str = "2021-10-30T03:12:00.000Z";
-        let sent_at = DateTime::parse_from_rfc3339(sent_at_str)
-            .unwrap()
-            .with_timezone(&Utc);
-
-        let result = parse_event_timestamp(Some(timestamp_str), None, Some(sent_at), true, now);
-
-        // Should use timestamp directly, ignoring sent_at
-        let expected = DateTime::parse_from_rfc3339(timestamp_str)
-            .unwrap()
-            .with_timezone(&Utc);
-        assert_eq!(result, expected);
-    }
-
-    #[test]
-    fn test_timezone_info_handling() {
-        // Matches: "captures sent_at with timezone info"
-        let now_str = "2021-10-29T01:44:00.000Z";
-        let now = DateTime::parse_from_rfc3339(now_str)
-            .unwrap()
-            .with_timezone(&Utc);
-        let timestamp_str = "2021-10-30T03:02:00.000+04:00"; // +04:00 timezone
-        let sent_at_str = "2021-10-30T03:12:00.000+04:00"; // Same timezone
-        let sent_at = DateTime::parse_from_rfc3339(sent_at_str)
-            .unwrap()
-            .with_timezone(&Utc);
-
-        let result = parse_event_timestamp(Some(timestamp_str), None, Some(sent_at), false, now);
-
-        // Should handle timezone conversion properly
-        // timestamp in UTC: 2021-10-29T23:02:00Z, sent_at in UTC: 2021-10-29T23:12:00Z
-        // Expected: now + (timestamp - sent_at) = 01:44:00 + (23:02:00 - 23:12:00) = 01:44:00 - 00:10:00 = 01:34:00
-        let expected = Utc.with_ymd_and_hms(2021, 10, 29, 1, 34, 0).unwrap();
-        assert_eq!(result, expected);
-    }
-
-    #[test]
-    fn test_timestamp_no_sent_at() {
-        // Matches: "captures timestamp with no sent_at"
-        let now_str = "2021-10-30T01:44:00.000Z";
-        let now = DateTime::parse_from_rfc3339(now_str)
-            .unwrap()
-            .with_timezone(&Utc);
-        let timestamp_str = "2021-10-30T03:02:00.000Z";
-
-        let result = parse_event_timestamp(Some(timestamp_str), None, None, false, now);
-
-        // Should use timestamp directly when no sent_at
-        let expected = DateTime::parse_from_rfc3339(timestamp_str)
-            .unwrap()
-            .with_timezone(&Utc);
-        assert_eq!(result, expected);
-    }
-
-    #[test]
-    fn test_offset_ignores_sent_at() {
-        // Matches: "captures with time offset and ignores sent_at"
-        let now_str = "2021-10-29T01:44:00.000Z";
-        let now = DateTime::parse_from_rfc3339(now_str)
-            .unwrap()
-            .with_timezone(&Utc);
-        let sent_at_str = "2021-10-30T03:12:00.000+04:00"; // Should be ignored
-        let sent_at = DateTime::parse_from_rfc3339(sent_at_str)
-            .unwrap()
-            .with_timezone(&Utc);
-        let offset = 6000; // 6 seconds
-
-        let result = parse_event_timestamp(None, Some(offset), Some(sent_at), false, now);
-
-        // Expected: now - offset = 01:44:00 - 6s = 01:43:54
-        let expected = Utc.with_ymd_and_hms(2021, 10, 29, 1, 43, 54).unwrap();
-        assert_eq!(result, expected);
-    }
-
-    #[test]
-    fn test_offset_only() {
-        // Matches: "captures with time offset"
-        let now_str = "2021-10-29T01:44:00.000Z";
-        let now = DateTime::parse_from_rfc3339(now_str)
-            .unwrap()
-            .with_timezone(&Utc);
-        let offset = 6000; // 6 seconds
-
-        let result = parse_event_timestamp(None, Some(offset), None, false, now);
-
-        // Expected: now - offset = 01:44:00 - 6s = 01:43:54
-        let expected = Utc.with_ymd_and_hms(2021, 10, 29, 1, 43, 54).unwrap();
-        assert_eq!(result, expected);
-    }
-
-    #[test]
-    fn test_extreme_offset_out_of_bounds() {
-        // Matches: "timestamps adjusted way out of bounds are ignored"
-        let now_str = "2021-10-28T01:10:00.000Z";
-        let now = DateTime::parse_from_rfc3339(now_str)
-            .unwrap()
-            .with_timezone(&Utc);
-        let timestamp_str = "2021-10-28T01:00:00.000Z";
-        let sent_at_str = "2021-10-28T01:05:00.000Z";
-        let sent_at = DateTime::parse_from_rfc3339(sent_at_str)
-            .unwrap()
-            .with_timezone(&Utc);
-        let offset = 600000000000000i64; // Extremely large offset
-
-        let result =
-            parse_event_timestamp(Some(timestamp_str), Some(offset), Some(sent_at), false, now);
-
-        // The large offset creates an underflow, resulting in a very old timestamp
-        // Our implementation actually does generate a warning for out-of-bounds results!
-        println!("Result timestamp: {result}");
-        println!("Expected (now): {now}");
-
-        // The result should be epoch time due to underflow
-        assert_eq!(result.year(), 1970); // Epoch time
-    }
-
-    #[test]
-    fn test_unparseable_timestamp_fallback() {
-        // Matches: "reports timestamp parsing error and fallbacks to DateTime.utc"
-        let now_str = "2020-08-12T01:02:00.000Z";
-        let now = DateTime::parse_from_rfc3339(now_str)
-            .unwrap()
-            .with_timezone(&Utc);
-        let invalid_timestamp = "notISO";
-
-        let result = parse_event_timestamp(Some(invalid_timestamp), None, None, false, now);
-
-        // Should fall back to now when timestamp is unparseable
-        assert_eq!(result, now);
-        // Note: Our Rust implementation doesn't generate warnings for unparseable timestamps,
-        // it just silently falls back to now. This is different from Node.js behavior.
-        // The warnings are only generated for invalid sent_at or out-of-bounds results.
-    }
-
-    #[test]
-    fn test_future_timestamp_with_sent_at_warning() {
-        // Matches: "reports event_timestamp_in_future with sent_at"
-        let now_str = "2021-10-29T01:00:00.000Z";
-        let now = DateTime::parse_from_rfc3339(now_str)
-            .unwrap()
-            .with_timezone(&Utc);
-        let timestamp_str = "2021-10-29T02:30:00.000Z"; // 1.5 hours in future
-        let sent_at_str = "2021-10-28T01:00:00.000Z"; // 24 hours ago
-        let sent_at = DateTime::parse_from_rfc3339(sent_at_str)
-            .unwrap()
-            .with_timezone(&Utc);
-
-        let result = parse_event_timestamp(Some(timestamp_str), None, Some(sent_at), false, now);
-
-        // Should clamp to now and generate warning
-        assert_eq!(result, now);
-    }
-
-    #[test]
-    fn test_future_timestamp_ignore_sent_at_warning() {
-        // Matches: "reports event_timestamp_in_future with $ignore_sent_at"
-        let now_str = "2021-09-29T01:00:00.000Z";
-        let now = DateTime::parse_from_rfc3339(now_str)
-            .unwrap()
-            .with_timezone(&Utc);
-        let timestamp_str = "2021-10-29T02:30:00.000Z"; // 30+ days in future
-
-        let result = parse_event_timestamp(Some(timestamp_str), None, None, true, now);
-
-        // Should clamp to now and generate warning
-        assert_eq!(result, now);
-    }
-
-    #[test]
-    fn test_future_timestamp_negative_offset_warning() {
-        // Matches: "reports event_timestamp_in_future with negative offset"
-        let now_str = "2021-10-29T01:00:00.000Z";
-        let now = DateTime::parse_from_rfc3339(now_str)
-            .unwrap()
-            .with_timezone(&Utc);
-        let offset = -82860000i64; // Large negative offset (creates future timestamp)
-
-        let result = parse_event_timestamp(None, Some(offset), None, false, now);
-
-        // Should clamp to now and generate warning
-        assert_eq!(result, now);
-    }
-
-    #[test]
-    fn test_iso8601_format_variants() {
-        // Based on Node.js parseDate tests - various ISO 8601 formats
-        let test_cases = vec![
-            "2021-10-29",
-            "2021-10-29 00:00:00",
-            "2021-10-29 00:00:00.000000",
-            "2021-10-29T00:00:00.000Z",
-            "2021-10-29 00:00:00+00:00",
-            "2021-10-29T00:00:00.000-00:00",
-            "2021-10-29T00:00:00.000",
-            "2021-10-29T00:00:00.000+00:00",
-        ];
-
-        for timestamp_str in test_cases {
-            let result = parse_date(timestamp_str);
-            assert!(
-                result.is_some(),
-                "Failed to parse ISO 8601 variant: {timestamp_str}"
-            );
-
-            let dt = result.unwrap();
-            println!(
-                "Parsed '{}' -> year: {}, month: {}, day: {}",
-                timestamp_str,
-                dt.year(),
-                dt.month(),
-                dt.day()
-            );
-            assert_eq!(dt.year(), 2021, "Wrong year for: {timestamp_str}");
-            assert_eq!(dt.month(), 10, "Wrong month for: {timestamp_str}");
-            // Some formats might be interpreted differently by dateparser
-            // Let's be more flexible and just ensure we get a valid October 2021 date
-            assert!(
-                dt.day() >= 28 && dt.day() <= 29,
-                "Wrong day for: {} (got {})",
-                timestamp_str,
-                dt.day()
-            );
-        }
-    }
-
-    #[test]
-    fn test_dateparser_timezone_conversion() {
-        // Test that dateparser correctly converts timezones to UTC
-        let test_cases = vec![
-            ("2023-01-01T12:00:00Z", 12, 0, 0),      // UTC
-            ("2023-01-01T12:00:00+00:00", 12, 0, 0), // UTC explicit
-            ("2023-01-01T12:00:00+03:00", 9, 0, 0),  // +3 hours should become 9 UTC
-            ("2023-01-01T12:00:00-05:00", 17, 0, 0), // -5 hours should become 17 UTC
-            ("2023-01-01T12:00:00+01:00", 11, 0, 0), // +1 hour should become 11 UTC
-            // Test the problematic format from your example
-            ("2025-09-17T14:05:04.805+03", 11, 5, 4), // +3 hours, should become 11:05:04 UTC
-        ];
-
-        for (input, expected_hour, expected_min, expected_sec) in test_cases {
-            let result = parse_date(input);
-            assert!(result.is_some(), "Failed to parse: {input}");
-
-            let dt = result.unwrap();
-            println!(
-                "Input: {} -> UTC: {}",
-                input,
-                dt.format("%Y-%m-%dT%H:%M:%S%.3fZ")
-            );
-
-            assert_eq!(
-                dt.hour(),
-                expected_hour,
-                "Wrong hour for {}: expected {}, got {}",
-                input,
-                expected_hour,
-                dt.hour()
-            );
-            assert_eq!(dt.minute(), expected_min, "Wrong minute for {input}");
-            assert_eq!(dt.second(), expected_sec, "Wrong second for {input}");
-        }
-    }
-
-    #[test]
-    fn test_normalize_timezone_format() {
-        // Test the timezone normalization function
-        let test_cases = vec![
-            // Cases that should be normalized
-            (
-                "2025-09-17T14:05:04.805+03",
-                "2025-09-17T14:05:04.805+03:00",
-            ),
-            (
-                "2025-09-17T14:05:04.805-05",
-                "2025-09-17T14:05:04.805-05:00",
-            ),
-            (
-                "2025-09-17T14:05:04.805+00",
-                "2025-09-17T14:05:04.805+00:00",
-            ),
-            ("2025-09-17T14:05:04+03", "2025-09-17T14:05:04+03:00"), // Without fractional seconds
-            ("2025-09-17T14:05:04-05", "2025-09-17T14:05:04-05:00"), // Without fractional seconds
-            // Cases that should NOT be normalized
-            ("2025-09-17T14:05:04.805Z", "2025-09-17T14:05:04.805Z"), // Already standard format
-            (
-                "2025-09-17T14:05:04.805+03:00",
-                "2025-09-17T14:05:04.805+03:00",
-            ), // Already standard format
-            ("2023-01-01", "2023-01-01"), // Date-only, should not be processed
-            ("2021-10-29", "2021-10-29"), // Date-only, should not be processed
-            ("invalid", "invalid"),       // No change for invalid input
-            ("2025-09-17T14:05:04.805+123", "2025-09-17T14:05:04.805+123"), // Invalid timezone (3 digits)
-            ("not-a-date-01", "not-a-date-01"), // Ends with -01 but not a datetime
-            ("T14:05:04+03", "T14:05:04+03"),   // Missing date part
-        ];
-
-        for (input, expected) in test_cases {
-            let result = normalize_timezone_format(input);
-            assert_eq!(result.as_ref(), expected, "Failed for input: {input}");
-        }
-    }
-
-    #[test]
-    fn test_advanced_iso8601_formats_rejected() {
-        // These advanced ISO 8601 formats are supported in Node.js but not in our Rust implementation
-        // Explicitly test that they are rejected to document the behavioral difference
-        let unsupported_cases = vec![
-            "2021-W43-5", // ISO week date format (week 43, day 5 = Friday = 2021-10-29)
-            "2021-302",   // ISO ordinal date format (day 302 of 2021 = 2021-10-29)
-        ];
-
-        for timestamp_str in unsupported_cases {
-            let result = parse_date(timestamp_str);
-            assert!(
-                result.is_none(),
-                "Expected advanced ISO format to be rejected: {timestamp_str}"
-            );
-            println!(
-                "✅ Advanced ISO format '{timestamp_str}' correctly rejected by Rust implementation"
-            );
-        }
-    }
-}
